+++ docs/content/ActionList.mdx
@@ -444,6 +444,6 @@ render(<SelectFields />)
 
 ## Related components
 
-- [ActionMenu](/drafts/ActionMenu2)
+- [ActionMenu](/ActionMenu)
 - [DropdownMenu](/DropdownMenu)
 - [SelectPanel](/SelectPanel)
