+++ docs/content/drafts/DropdownMenu2.mdx
@@ -360,5 +360,5 @@ Use `DropdownMenu` to select an option from a small list. If you’re looking fo
 ## Related components
 
 - [ActionList](/ActionList)
-- [ActionMenu](/ActionMenu2)
+- [ActionMenu](/ActionMenu)
 - [SelectPanel](/SelectPanel)
