+++ docs/src/@primer/gatsby-theme-doctocat/live-code-scope.js
@@ -18,7 +18,7 @@ import {
   MarkGithubIcon,
   NoteIcon,
   NumberIcon,
-  OctofaceIcon,
+  SmileyIcon,
   PencilIcon,
   PersonIcon,
   ProjectIcon,
@@ -60,7 +60,7 @@ export default function resolveScope(metastring) {
     ZapIcon,
     XIcon,
     DotIcon,
-    OctofaceIcon,
+    SmileyIcon,
     PersonIcon,
     MailIcon,
     GitCommitIcon,
