+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -37,6 +37,8 @@
   children:
     - title: ActionList
       url: /ActionList
+    - title: ActionMenu
+      url: /ActionMenu
     - title: Autocomplete
       url: /Autocomplete
     - title: Avatar
@@ -145,25 +147,25 @@
       url: /drafts/LinkButton
     - title: IconButton
       url: /drafts/IconButton
-    - title: ActionMenu v2
-      url: /drafts/ActionMenu2
 - title: Deprecated
   children:
+    - title: ActionList
+      url: /deprecated/ActionList
+    - title: ActionMenu
+      url: /deprecated/ActionMenu
     - title: BorderBox
       url: /deprecated/BorderBox
-    - title: Flex
-      url: /deprecated/Flex
-    - title: Grid
-      url: /deprecated/Grid
-    - title: Position
-      url: /deprecated/Position
     - title: Dialog
       url: /deprecated/Dialog
     - title: Dropdown
       url: /deprecated/Dropdown
+    - title: Flex
+      url: /deprecated/Flex
     - title: FormGroup
       url: /FormGroup
+    - title: Grid
+      url: /deprecated/Grid
+    - title: Position
+      url: /deprecated/Position
     - title: SelectMenu
       url: /deprecated/SelectMenu
-    - title: ActionList
-      url: /deprecated/ActionList
