+++ src/ActionMenu.tsx
@@ -1,106 +1,131 @@
-import {GroupedListProps, List, ListPropsBase} from './deprecated/ActionList/List'
-import {Item, ItemProps} from './deprecated/ActionList/Item'
-import {Divider} from './deprecated/ActionList/Divider'
-import Button, {ButtonProps} from './Button'
-import React, {useCallback, useMemo} from 'react'
-import {AnchoredOverlay} from './AnchoredOverlay'
-import {useProvidedStateOrCreate} from './hooks/useProvidedStateOrCreate'
+import React from 'react'
+import {useSSRSafeId} from '@react-aria/ssr'
+import {TriangleDownIcon} from '@primer/octicons-react'
+import {AnchoredOverlay, AnchoredOverlayProps} from './AnchoredOverlay'
 import {OverlayProps} from './Overlay'
-import {useProvidedRefOrCreate} from './hooks'
-import {AnchoredOverlayWrapperAnchorProps} from './AnchoredOverlay/AnchoredOverlay'
+import {useProvidedRefOrCreate, useProvidedStateOrCreate, useMenuInitialFocus, useTypeaheadFocus} from './hooks'
+import {Divider} from './ActionList/Divider'
+import {ActionListContainerContext} from './ActionList/ActionListContainerContext'
+import {Button, ButtonProps} from './Button2'
+import {MandateProps} from './utils/types'
 
-interface ActionMenuBaseProps extends Partial<Omit<GroupedListProps, keyof ListPropsBase>>, ListPropsBase {
+type MenuContextProps = Pick<
+  AnchoredOverlayProps,
+  'anchorRef' | 'renderAnchor' | 'open' | 'onOpen' | 'onClose' | 'anchorId'
+>
+const MenuContext = React.createContext<MenuContextProps>({renderAnchor: null, open: false})
+
+export type ActionMenuProps = {
   /**
-   * Content that is passed into the renderAnchor component, which is a button by default.
+   * Recommended: `ActionMenu.Button` or `ActionMenu.Anchor` with `ActionMenu.Overlay`
    */
-  anchorContent?: React.ReactNode
+  children: React.ReactElement[] | React.ReactElement
 
   /**
-   * A callback that triggers both on clicks and keyboard events. This callback will be overridden by item level `onAction` callbacks.
-   */
-  onAction?: (props: ItemProps, event?: React.MouseEvent<HTMLDivElement> | React.KeyboardEvent<HTMLDivElement>) => void
-
-  /**
-   * If defined, will control the open/closed state of the overlay. Must be used in conjuction with `setOpen`.
+   * If defined, will control the open/closed state of the overlay. Must be used in conjuction with `onOpenChange`.
    */
   open?: boolean
 
   /**
    * If defined, will control the open/closed state of the overlay. Must be used in conjuction with `open`.
    */
-  setOpen?: (s: boolean) => void
+  onOpenChange?: (s: boolean) => void
+} & Pick<AnchoredOverlayProps, 'anchorRef'>
 
-  /**
-   * Props to be spread on the internal `Overlay` component.
-   */
-  overlayProps?: Partial<OverlayProps>
-}
-
-export type ActionMenuProps = ActionMenuBaseProps & AnchoredOverlayWrapperAnchorProps
-
-const ActionMenuItem = (props: ItemProps) => <Item role="menuitem" {...props} />
-
-ActionMenuItem.displayName = 'ActionMenu.Item'
-
-const ActionMenuBase = ({
-  anchorContent,
-  renderAnchor = <T extends ButtonProps>(props: T) => <Button {...props} />,
+const Menu: React.FC<ActionMenuProps> = ({
   anchorRef: externalAnchorRef,
-  onAction,
   open,
-  setOpen,
-  overlayProps,
-  items,
-  ...listProps
-}: ActionMenuProps): JSX.Element => {
-  const [combinedOpenState, setCombinedOpenState] = useProvidedStateOrCreate(open, setOpen, false)
-  const anchorRef = useProvidedRefOrCreate(externalAnchorRef)
-  const onOpen = useCallback(() => setCombinedOpenState(true), [setCombinedOpenState])
-  const onClose = useCallback(() => setCombinedOpenState(false), [setCombinedOpenState])
+  onOpenChange,
+  children
+}: ActionMenuProps) => {
+  const [combinedOpenState, setCombinedOpenState] = useProvidedStateOrCreate(open, onOpenChange, false)
+  const onOpen = React.useCallback(() => setCombinedOpenState(true), [setCombinedOpenState])
+  const onClose = React.useCallback(() => setCombinedOpenState(false), [setCombinedOpenState])
 
-  const renderMenuAnchor = useMemo(() => {
-    if (renderAnchor === null) {
+  const anchorRef = useProvidedRefOrCreate(externalAnchorRef)
+  const anchorId = useSSRSafeId()
+  let renderAnchor: AnchoredOverlayProps['renderAnchor'] = null
+
+  // 🚨 Hack for good API!
+  // we strip out Anchor from children and pass it to AnchoredOverlay to render
+  // with additional props for accessibility
+  const contents = React.Children.map(children, child => {
+    if (child.type === MenuButton || child.type === Anchor) {
+      renderAnchor = anchorProps => React.cloneElement(child, anchorProps)
       return null
     }
-    return <T extends React.HTMLAttributes<HTMLElement>>(props: T) => {
-      return renderAnchor({
-        'aria-label': 'menu',
-        children: anchorContent,
-        ...props
-      })
-    }
-  }, [anchorContent, renderAnchor])
+    return child
+  })
 
-  const itemsToRender = useMemo(() => {
-    return items.map(item => {
-      return {
-        ...item,
-        role: 'menuitem',
-        onAction: (props, event) => {
-          const actionCallback = item.onAction ?? onAction
-          actionCallback?.(props as ItemProps, event)
-          if (!event.defaultPrevented) {
-            onClose()
-          }
-        }
-      } as ItemProps
-    })
-  }, [items, onAction, onClose])
+  return (
+    <MenuContext.Provider value={{anchorRef, renderAnchor, anchorId, open: combinedOpenState, onOpen, onClose}}>
+      {contents}
+    </MenuContext.Provider>
+  )
+}
+
+export type ActionMenuAnchorProps = {children: React.ReactElement}
+const Anchor = React.forwardRef<AnchoredOverlayProps['anchorRef'], ActionMenuAnchorProps>(
+  ({children, ...anchorProps}, anchorRef) => {
+    return React.cloneElement(children, {...anchorProps, ref: anchorRef})
+  }
+)
+
+/** this component is syntactical sugar 🍭 */
+export type ActionMenuButtonProps = ButtonProps
+const MenuButton = React.forwardRef<AnchoredOverlayProps['anchorRef'], ButtonProps>((props, anchorRef) => {
+  return (
+    <Anchor ref={anchorRef}>
+      <Button trailingIcon={TriangleDownIcon} type="button" {...props} />
+    </Anchor>
+  )
+})
+
+type MenuOverlayProps = Partial<OverlayProps> & {
+  /**
+   * Recommended: `ActionList`
+   */
+  children: React.ReactElement[] | React.ReactElement
+}
+const Overlay: React.FC<MenuOverlayProps> = ({children, ...overlayProps}) => {
+  // we typecast anchorRef as required instead of optional
+  // because we know that we're setting it in context in Menu
+  const {anchorRef, renderAnchor, anchorId, open, onOpen, onClose} = React.useContext(MenuContext) as MandateProps<
+    MenuContextProps,
+    'anchorRef'
+  >
+
+  const containerRef = React.createRef<HTMLDivElement>()
+  const {openWithFocus} = useMenuInitialFocus(open, onOpen, containerRef)
+  useTypeaheadFocus(open, containerRef)
 
   return (
     <AnchoredOverlay
-      renderAnchor={renderMenuAnchor}
       anchorRef={anchorRef}
-      open={combinedOpenState}
-      onOpen={onOpen}
+      renderAnchor={renderAnchor}
+      anchorId={anchorId}
+      open={open}
+      onOpen={openWithFocus}
       onClose={onClose}
       overlayProps={overlayProps}
+      focusZoneSettings={{focusOutBehavior: 'wrap'}}
     >
-      <List {...listProps} role="menu" items={itemsToRender} />
+      <div ref={containerRef}>
+        <ActionListContainerContext.Provider
+          value={{
+            container: 'ActionMenu',
+            listRole: 'menu',
+            listLabelledBy: anchorId,
+            selectionAttribute: 'aria-checked', // Should this be here?
+            afterSelect: onClose
+          }}
+        >
+          {children}
+        </ActionListContainerContext.Provider>
+      </div>
     </AnchoredOverlay>
   )
 }
 
-ActionMenuBase.displayName = 'ActionMenu'
-
-export const ActionMenu = Object.assign(ActionMenuBase, {Divider, Item: ActionMenuItem})
+Menu.displayName = 'ActionMenu'
+export const ActionMenu = Object.assign(Menu, {Button: MenuButton, Anchor, Overlay, Divider})
