+++ src/deprecated/index.ts
@@ -39,3 +39,5 @@ export type {
 } from '../SelectMenu'
 export {ActionList} from './ActionList'
 export type {ActionListProps} from './ActionList'
+export {ActionMenu} from './ActionMenu'
+export type {ActionMenuProps} from './ActionMenu'
