+++ src/drafts/index.ts
@@ -7,6 +7,5 @@
 
 // Components
 export * from '../Button2'
-export * from '../ActionMenu2'
 export * from '../DropdownMenu2'
 export * from '../Label2'
