+++ src/index.ts
@@ -39,7 +39,7 @@ export type {
   ActionListTrailingVisualProps
 } from './ActionList'
 export {ActionMenu} from './ActionMenu'
-export type {ActionMenuProps} from './ActionMenu'
+export type {ActionMenuProps, ActionMenuAnchorProps, ActionMenuButtonProps} from './ActionMenu'
 export {default as Autocomplete} from './Autocomplete'
 export type {AutocompleteMenuProps, AutocompleteInputProps, AutocompleteOverlayProps} from './Autocomplete'
 export {default as Avatar} from './Avatar'
